Refactor sync group creation to player manager
