+++ music_assistant/common/models/enums.py
@@ -370,7 +370,6 @@ class ProviderFeature(StrEnum):
     #
     # PLAYERPROVIDER FEATURES
     #
-    PLAYER_GROUP_CREATE = "player_group_create"
     SYNC_PLAYERS = "sync_players"
 
     #
