+++ music_assistant/server/providers/snapcast/__init__.py
@@ -246,7 +246,7 @@ class SnapCastProvider(PlayerProvider):
     @property
     def supported_features(self) -> tuple[ProviderFeature, ...]:
         """Return the features supported by this Provider."""
-        return (ProviderFeature.SYNC_PLAYERS, ProviderFeature.PLAYER_GROUP_CREATE)
+        return (ProviderFeature.SYNC_PLAYERS, ProviderFeature.SYNC_GROUP)
 
     async def handle_async_init(self) -> None:
         """Handle async initialization of the provider."""
