+++ music_assistant/server/providers/sonos_s1/__init__.py
@@ -137,7 +137,7 @@ class SonosPlayerProvider(PlayerProvider):
     @property
     def supported_features(self) -> tuple[ProviderFeature, ...]:
         """Return the features supported by this Provider."""
-        return (ProviderFeature.SYNC_PLAYERS, ProviderFeature.PLAYER_GROUP_CREATE)
+        return (ProviderFeature.SYNC_PLAYERS, ProviderFeature.SYNC_GROUP)
 
     async def handle_async_init(self) -> None:
         """Handle async initialization of the provider."""
